Replaced the original Simulation class with the SimulationRewrite class.
