+++ agents_playground/core/simulation.py
@@ -1,22 +1,28 @@
-from abc import ABC, abstractmethod
+"""
+Single file rewrite of coroutine based simulation.
+Prototyping the class design. Will break into modules if this pans out.
+"""
+from __future__ import annotations
 from collections import OrderedDict
-
-from math import floor
-import threading
-from time import sleep
-from typing import Callable, Dict, Optional,  Union
+import os
+from types import SimpleNamespace
+from typing import Callable, NamedTuple, Optional, Union
 
 import dearpygui.dearpygui as dpg
 
 from agents_playground.core.observe import Observable
-from agents_playground.core.time_utilities import (
-  MS_PER_SEC,
-  TIME_PER_FRAME,
-  UPDATE_BUDGET, 
-  TimeInMS, 
-  TimeUtilities
-)
+from agents_playground.core.sim_loop import SimLoop
+from agents_playground.core.task_scheduler import TaskScheduler
 from agents_playground.core.callable_utils import CallableUtility
+from agents_playground.entities.entities_registry import ENTITIES_REGISTRY
+from agents_playground.renderers.entities import render_entities
+from agents_playground.renderers.renderers_registry import RENDERERS_REGISTRY
+from agents_playground.renderers.agent import render_agents_scene
+from agents_playground.renderers.grid import render_grid
+from agents_playground.renderers.path import render_interpolated_paths
+from agents_playground.renderers.stats import render_stats
+from agents_playground.scene.scene_builder import SceneBuilder
+from agents_playground.sims.multiple_agents_sim import agents_spinning
 from agents_playground.simulation.context import SimulationContext
 from agents_playground.simulation.render_layer import RenderLayer
 from agents_playground.simulation.sim_events import SimulationEvents
@@ -25,45 +31,102 @@ from agents_playground.simulation.sim_state import (
   SimulationStateTable,
   SimulationStateToLabelMap
 )
-from agents_playground.simulation.statistics import SimulationStatistics
 from agents_playground.simulation.tag import Tag
 from agents_playground.sys.logger import get_default_logger
+from agents_playground.scene.scene_reader import SceneReader
+from agents_playground.tasks.tasks_registry import TASKS_REGISTRY
 
 logger = get_default_logger()
 
-class Simulation(ABC, Observable):
+class SimulationUIComponents(NamedTuple):
+  sim_window_ref: Tag
+  sim_menu_bar_ref: Tag
+  sim_initial_state_dl_ref: Tag
+  buttons: dict
+
+"""
+https://app.diagrams.net/?src=about#G1UgUgEILyWFRzwlZu0GIsVdS7WhzqWH3G
+Current class stats:
+- fields: 18 -> 10
+- Properties: 3 -> 2
+- Public Methods: 2 -> 2
+- Private Methods: 20 -> 11
+
+Possible Refactors
+- Make SimLoop something that can be passed in to enable other looping types.
+  How to make this more generic? 
+    - The update_statistics is very specific. 
+    - I'd like to be able to see a chart of the utilization over time.
+    - The task_scheduler is passed in.
+- Should _establish_context be 100% in TOML files. Perhaps there should be a 
+  default TOML file to keep things like font, that can be overridden?
+
+
+TODO
+- Check Typing
+- Unit Tests
+- Diagram
+- Update the module imports
+- PyDoc
+- Blog
+- Utility Method style simulation
+- Maze port the maze sim to use TOML and other AI techniques?
+"""
+class Simulation(Observable):
+  """This class may potentially replace Simulation."""
   _primary_window_ref: Union[int, str]
 
-  def __init__(self) -> None:
+  def __init__(self, scene_toml: str) -> None:
     super().__init__()
     logger.info('Simulation: Initializing')
-    self._sim_current_state: SimulationState = SimulationState.INITIAL
+    self._scene_toml = scene_toml
     self._context: SimulationContext = SimulationContext(dpg.generate_uuid)
-    self._sim_window_ref = dpg.generate_uuid()
-    self._sim_menu_bar_ref = dpg.generate_uuid()
-    self._sim_initial_state_dl_ref = dpg.generate_uuid()
-    self._buttons = {
-      'sim': {
-        'run_sim_toggle_btn': dpg.generate_uuid()
+    self._ui_components = SimulationUIComponents(
+      dpg.generate_uuid(), 
+      dpg.generate_uuid(), 
+      dpg.generate_uuid(), 
+      {
+        'sim': {
+          'run_sim_toggle_btn': dpg.generate_uuid()
+        }
       }
-    }
+    )
     self._layers: OrderedDict[Tag, RenderLayer] = OrderedDict()
-    self._sim_run_rate: float = 0.200 #How fast to run the simulation.
     self._title: str = "Set the Simulation Title"
     self._sim_description = 'Set the Simulation Description'
     self._sim_instructions = 'Set the Simulation Instructions'
-    self._sim_stopped_check_time: float = 0.5
-    self._fps_rate: float = 0
-    self._utilization_rate: float = 0
-    self.add_layer(render_stats, "Statistics")
+    self._task_scheduler = TaskScheduler()
+    self._sim_loop = SimLoop(scheduler = self._task_scheduler)
+    
+    """
+    TODO: This appears to be short sighted. Needs to be configurable somehow.
+
+    I need a better way to register render functions to be called. 
+    The render_interpolated_paths is looping through the list of paths and 
+    calling the associated render function. Where in contrast the Agents don't 
+    have a render method, there is just the centralized render_agents_scene function
+    that renders all the agents in a single batch. 
+
+    The render_interpolated_paths approach seems more flexible. The simplest 
+    thing I can think of is to have a Renderable interface that exposes a render() 
+    method and a renderer field. Then things that can be rendered simply have 
+    implement that interface and add themselves to a list of renderable entities.
+
+    I need a way to void adding a one off function for every new simulation use case.
+    """
+    self.add_layer(render_stats, 'Statistics')
+    self.add_layer(render_grid, 'Terrain')
+    self.add_layer(render_entities, 'Entities')
+    self.add_layer(render_interpolated_paths, 'Path')
+    self.add_layer(render_agents_scene, 'Agents') 
 
   @property
   def simulation_state(self) -> SimulationState:
-    return self._sim_current_state
+    return self._sim_loop.simulation_state
 
   @simulation_state.setter
   def simulation_state(self, next_state: SimulationState) -> None:
-    self._sim_current_state = next_state
+    self._sim_loop.simulation_state = next_state
 
   @property
   def primary_window(self) -> Union[int, str]:
@@ -74,14 +137,6 @@ class Simulation(ABC, Observable):
   def primary_window(self, primary_window_ref: Union[int, str]) -> None:
     """Assigns the primary window to the simulation window."""
     self._primary_window_ref = primary_window_ref
-
-  @property
-  def simulation_title(self) -> str:
-    return self._title
-
-  @simulation_title.setter
-  def simulation_title(self, value: str) -> None:
-    self._title = value
 
   def add_layer(self, layer: Callable, label: str) -> None:
     """Adds a layer
@@ -100,35 +155,48 @@ class Simulation(ABC, Observable):
   def launch(self):
     """Opens the Simulation Window"""
     logger.info('Simulation: Launching')
+    self._load_scene()
     parent_width: Optional[int] = dpg.get_item_width(self.primary_window)
     parent_height: Optional[int] = dpg.get_item_height(self.primary_window)
 
-    with dpg.window(tag=self._sim_window_ref, 
-      label=self.simulation_title, 
+    with dpg.window(tag=self._ui_components.sim_window_ref, 
+      label=self._title, 
       width=parent_width, 
       height=parent_height, 
       on_close=self._handle_sim_closed):
       self._setup_menu_bar()
-      if self._sim_current_state is SimulationState.INITIAL:
+      if self._sim_loop.simulation_state is SimulationState.INITIAL:
         self._initial_render()
       else:
         self._start_simulation()
+
+  def _load_scene(self):
+    """Load the scene data from a TOML file."""
+    logger.info('Simulation: Loading Scene')
+    scene_reader = SceneReader()
+    scene_path = os.path.abspath(self._scene_toml)
+    scene_data:SimpleNamespace = scene_reader.load(scene_path)
+
+    # Setup UI
+    self._title = scene_data.simulation.ui.title
+    self._sim_description = scene_data.simulation.ui.description
+    self._sim_instructions = scene_data.simulation.ui.instructions
+
+    scene_builder = SceneBuilder(
+      id_generator = dpg.generate_uuid, 
+      task_scheduler = self._task_scheduler,
+      render_map = RENDERERS_REGISTRY, 
+      task_map = TASKS_REGISTRY,
+      entities_map = ENTITIES_REGISTRY
+    )
+
+    self._context.scene = scene_builder.build(scene_data)
 
   def _start_simulation(self):
     logger.info('Simulation: Starting simulation')
     self._establish_context()
     self._initialize_layers()
-    self._bootstrap_simulation_render()
-
-    # Create a thread for updating the simulation.
-    # Note: A daemonic thread cannot be "joined" by another thread. 
-    # They are destroyed when the main thread is terminated.
-    self._sim_thread = threading.Thread( #name="single-agent-thread", 
-      target=self._sim_loop, 
-      args=(), 
-      daemon=True
-    )
-    self._sim_thread.start()
+    self._sim_loop.start(self._context)
 
   def _establish_context(self) -> None:
     '''Setups the variables used by the simulation.'''
@@ -142,13 +210,12 @@ class Simulation(ABC, Observable):
     self._context.agent_style.fill_color = (0, 0, 255)
     self._context.agent_style.size.width = 20
     self._context.agent_style.size.height = 20
-    self._establish_context_ext(self._context)
 
   def _initialize_layers(self) -> None:
     """Initializes the rendering code for each registered layer."""
     logger.info('Simulation: Initializing Layers')
     with dpg.drawlist(
-      parent=self._sim_window_ref, 
+      parent=self._ui_components.sim_window_ref, 
       width=self._context.canvas.width, 
       height=self._context.canvas.height):
       for rl in self._layers.values():
@@ -165,10 +232,13 @@ class Simulation(ABC, Observable):
 
   def _setup_menu_bar(self):
     logger.info('Simulation: Setting up the menu bar.')
-    with dpg.menu_bar(tag=self._sim_menu_bar_ref):
-      dpg.add_button(label=SimulationStateToLabelMap[self._sim_current_state], tag=self._buttons['sim']['run_sim_toggle_btn'], callback=self._run_sim_toggle_btn_clicked)
+    with dpg.menu_bar(tag=self._ui_components.sim_menu_bar_ref):
+      dpg.add_button(
+        label=SimulationStateToLabelMap[self._sim_loop.simulation_state], 
+        tag=self._ui_components.buttons['sim']['run_sim_toggle_btn'], 
+        callback=self._run_sim_toggle_btn_clicked
+      )
       self._setup_layers_menu()
-      self._setup_menu_bar_ext()
 
   def _setup_layers_menu(self) -> None:
     logger.info('Simulation: Setting up layer\'s menu.')
@@ -196,62 +266,10 @@ class Simulation(ABC, Observable):
   
     if self.simulation_state is SimulationState.INITIAL:
     # special case for starting the simulation for the first time.
-      if dpg.does_item_exist(self._sim_initial_state_dl_ref):
-        dpg.delete_item(self._sim_initial_state_dl_ref) 
+      if dpg.does_item_exist(self._ui_components.sim_initial_state_dl_ref):
+        dpg.delete_item(self._ui_components.sim_initial_state_dl_ref) 
       self._start_simulation()
   
-  def _sim_loop(self, **data):
-    """The thread callback that processes a simulation tick.
-    
-    Using the definitions in agents_playground.core.time_utilities, this enforces 
-    a fixed time step of TIME_PER_FRAME assuming a constant frame rate of 
-    TARGET_FRAMES_PER_SEC.
-
-    It prevents a simulation from running faster than the target frame rate but 
-    does nothing to help during a slow down.
-
-    The function determines the current frame in a 1 second rolling window and passes
-    that to the _sim_loop_tick() as the tick parameter.
-    """
-    data['current_second_start'] = 0.0 # TimeInMS
-    data['current_second_end'] = data['current_second_start'] + MS_PER_SEC # TimeInMS
-    while self.simulation_state is not SimulationState.ENDED:
-      if self.simulation_state is SimulationState.RUNNING:
-        self._process_sim_cycle(**data)
-      else:
-        # Give the CPU a break and sleep a bit before checking if we're still paused.
-        sleep(self._sim_stopped_check_time) 
-
-  def _process_sim_cycle(self, **data) -> None:
-    frame_start: TimeInMS = TimeUtilities.now()
-    loop_stats = {}
-    if (frame_start > data['current_second_end']):
-      # We've started a new second. 
-      data['current_second_start'] = frame_start
-      data['current_second_end'] = data['current_second_start'] + MS_PER_SEC
-    frame = floor((frame_start - data['current_second_start'])/TIME_PER_FRAME) + 1
-    loop_stats['time_started_running_tasks'] = TimeUtilities.now()
-    self._sim_loop_tick(current_frame=frame)
-    loop_stats['time_finished_running_tasks'] = TimeUtilities.now()
-    self._update_statistics(loop_stats)
-    amount_to_sleep_ms = frame_start + TIME_PER_FRAME - TimeUtilities.now()
-    amount_to_sleep_sec = amount_to_sleep_ms/MS_PER_SEC
-    if (amount_to_sleep_ms > 0):
-      sleep(amount_to_sleep_sec) 
-
-  # TODO Split the stats calculations from the rendering. 
-  # Probably only need to render once a second, not every cycle.
-  def _update_statistics(self, stats: dict[str, float]) -> None:
-    self._context.stats.fps.value = dpg.get_frame_rate()
-    self._context.stats.utilization.value = round(((stats['time_finished_running_tasks'] - stats['time_started_running_tasks'])/UPDATE_BUDGET) * 100, 2)
-
-    # This is will cause a render. Need to be smart with how these are grouped.
-    # There may be a way to do all the scene graph manipulation and configure_item
-    # calls in a single buffer.
-    # TODO Look at https://dearpygui.readthedocs.io/en/latest/documentation/staging.html
-    dpg.configure_item(item=self._context.stats.fps.id, text=f'Frame Rate (Hz): {self._context.stats.fps.value}')
-    dpg.configure_item(item=self._context.stats.utilization.id, text=f'Utilization (%): {self._context.stats.utilization.value}')  
-
   def _toggle_layer(self, sender, item_data, user_data):
     if user_data:
       if item_data:
@@ -266,35 +284,9 @@ class Simulation(ABC, Observable):
     canvas_width: int = parent_width if parent_width else 0
     canvas_height: int = parent_height - 40 if parent_height else 0
 
-    with dpg.drawlist(tag=self._sim_initial_state_dl_ref, parent=self._sim_window_ref, width=canvas_width, height=canvas_height): 
+    with dpg.drawlist(
+      tag=self._ui_components.sim_initial_state_dl_ref, 
+      parent=self._ui_components.sim_window_ref, 
+      width=canvas_width, height=canvas_height): 
       dpg.draw_text(pos=(20,20), text=self._sim_description, size=13)
       dpg.draw_text(pos=(20,40), text=self._sim_instructions, size=13)
-
-  @abstractmethod
-  def _bootstrap_simulation_render(self) -> None:
-    """Define the render setup for when the render is started."""
-
-  @abstractmethod
-  def _sim_loop_tick(self, **args):
-    """Handles one tick of the simulation."""
-
-  @abstractmethod
-  def _setup_menu_bar_ext(self) -> None:
-    """Setup simulation specific menu items."""
-  
-  @abstractmethod
-  def _establish_context_ext(self, context: SimulationContext) -> None:
-    """Setup simulation specific context variables."""
-
-# TODO: Find a home.
-def render_stats(**data) -> None:
-  """Render a text overlay of the active runtime statistics.
-  
-  Args:
-    - 
-  """
-  context: SimulationContext = data['context']
-  stats = context.stats
-  # TODO Need to make the stats text display on top of the terrain.
-  dpg.draw_text(tag=stats.fps.id, pos=(20,20), text=f'Frame Rate (Hz): {stats.fps.value}', size=13)
-  dpg.draw_text(tag=stats.utilization.id, pos=(20,40), text=f'Utilization (%): {stats.utilization.value}', size=13)
